Remove AppAdminController, add admin creation endpoint to BloodBankAdminController
The AppAdminController has been removed from the BloodBankBackEnd project. A new endpoint for creating a blood bank admin has been added to the BloodBankAdminController. The endpoint accepts a request containing the admin's information and creates a new admin record in the database. The frontend code for adding a hospital staff member has been updated to reflect these changes.